% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloatlas-package.R
\docType{package}
\name{phyloatlas-package}
\alias{phyloatlas}
\alias{phyloatlas-package}
\title{phyloatlas: R access to the Phylo-Species Atlas}
\description{
Convenience functions to fetch phylogenetic trees and metadata from the
Phylo-Species Atlas directly from R.
}
\details{
Key functions:
* [load_atlas_tree()] - load a tree by name, with species labels.
* [list_trees()] - data frame of all available trees and metadata.
* [atlas_info()] - metadata for a single tree.
* [atlas_clear_cache()] - drop the in-session metadata/dictionary cache.

The atlas lives at <https://github.com/franciscorichter/phylo-species-atlas>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franciscorichter/phylo-species-atlas}
  \item \url{https://franciscorichter.github.io/phylo-species-atlas/}
  \item Report bugs at \url{https://github.com/franciscorichter/phylo-species-atlas/issues}
}

}
\author{
\strong{Maintainer}: Francisco Richter \email{richtf@usi.ch} (\href{https://orcid.org/0000-0002-0924-4613}{ORCID})

}
\keyword{internal}
